# Optimize loading habits on Agenda Screen