--- a/core/data/src/debug/java/com/rendox/routinetracker/core/data/di/DatabasePrepopulatorModule.kt
+++ b/core/data/src/debug/java/com/rendox/routinetracker/core/data/di/DatabasePrepopulatorModule.kt
@@ -1,0 +1,16 @@
+package com.rendox.routinetracker.core.data.di
+
+import com.rendox.routinetracker.core.data.DatabasePrepopulator
+import com.rendox.routinetracker.core.data.DatabasePrepopulatorRandom
+import com.rendox.routinetracker.core.data.RandomHabitsGenerator
+import org.koin.dsl.module
+
+val databasePrepopulatorModule = module {
+    single<DatabasePrepopulator> {
+        DatabasePrepopulatorRandom(
+            habitsGenerator = RandomHabitsGenerator(numOfHabits = 100),
+            habitLocalDataSource = get(),
+            completionHistoryLocalDataSource = get(),
+        )
+    }
+}