--- a/core/data/src/main/java/com/rendox/routinetracker/core/data/DatabasePrepopulator.kt
+++ b/core/data/src/main/java/com/rendox/routinetracker/core/data/DatabasePrepopulator.kt
@@ -1,0 +1,5 @@
+package com.rendox.routinetracker.core.data
+
+interface DatabasePrepopulator {
+    suspend fun prepopulateDatabase()
+}