--- a/core/data/src/main/java/com/rendox/routinetracker/core/data/DatabasePrepopulatorNoOp.kt
+++ b/core/data/src/main/java/com/rendox/routinetracker/core/data/DatabasePrepopulatorNoOp.kt
@@ -1,0 +1,7 @@
+package com.rendox.routinetracker.core.data
+
+class DatabasePrepopulatorNoOp : DatabasePrepopulator {
+    override suspend fun prepopulateDatabase() {
+        // No-op
+    }
+}