--- a/core/data/src/main/java/com/rendox/routinetracker/core/data/completion_history/CompletionHistoryRepository.kt
+++ b/core/data/src/main/java/com/rendox/routinetracker/core/data/completion_history/CompletionHistoryRepository.kt
@@ -33,7 +33,7 @@ interface CompletionHistoryRepository {
         maxDate: LocalDate? = null,
     ): List<Habit.CompletionRecord>
 
-    suspend fun getAllRecords(): List<Pair<Long, Habit.CompletionRecord>>
+    suspend fun getAllRecords(): Map<Long, List<Habit.CompletionRecord>>
 
     suspend fun insertCompletion(
         habitId: Long,