--- a/core/data/src/main/java/com/rendox/routinetracker/core/data/completion_history/CompletionHistoryRepository.kt
+++ b/core/data/src/main/java/com/rendox/routinetracker/core/data/completion_history/CompletionHistoryRepository.kt
@@ -4,24 +4,6 @@ import com.rendox.routinetracker.core.model.Habit
 import kotlinx.datetime.LocalDate
 
 interface CompletionHistoryRepository {
-    suspend fun getNumOfTimesCompletedInPeriod(
-        habitId: Long,
-        minDate: LocalDate?,
-        maxDate: LocalDate?,
-    ): Double
-
-    suspend fun getRecordByDate(habitId: Long, date: LocalDate): Habit.CompletionRecord?
-    suspend fun getLastCompletedRecord(
-        habitId: Long,
-        minDate: LocalDate? = null,
-        maxDate: LocalDate? = null,
-    ): Habit.CompletionRecord?
-
-    suspend fun getFirstCompletedRecord(
-        habitId: Long,
-        minDate: LocalDate? = null,
-        maxDate: LocalDate? = null,
-    ): Habit.CompletionRecord?
 
     /**
      * get all records for the given habit in the ascending order, if [minDate] or [maxDate] are