--- a/core/data/src/main/java/com/rendox/routinetracker/core/data/completion_history/CompletionHistoryRepository.kt
+++ b/core/data/src/main/java/com/rendox/routinetracker/core/data/completion_history/CompletionHistoryRepository.kt
@@ -10,7 +10,7 @@ interface CompletionHistoryRepository {
      * specified, returns only the records in the given range.
      */
     suspend fun getRecordsInPeriod(
-        habitId: Long,
+        habit: Habit,
         minDate: LocalDate? = null,
         maxDate: LocalDate? = null,
     ): List<Habit.CompletionRecord>