--- a/core/data/src/main/java/com/rendox/routinetracker/core/data/completion_history/CompletionHistoryRepository.kt
+++ b/core/data/src/main/java/com/rendox/routinetracker/core/data/completion_history/CompletionHistoryRepository.kt
@@ -1,5 +1,6 @@
 package com.rendox.routinetracker.core.data.completion_history
 
+import com.rendox.routinetracker.core.logic.time.LocalDateRange
 import com.rendox.routinetracker.core.model.Habit
 import kotlinx.datetime.LocalDate
 
@@ -15,7 +16,9 @@ interface CompletionHistoryRepository {
         maxDate: LocalDate? = null,
     ): List<Habit.CompletionRecord>
 
-    suspend fun getAllRecords(): Map<Long, List<Habit.CompletionRecord>>
+    suspend fun getMultiHabitRecords(
+        habitsToPeriods: List<Pair<List<Habit>, LocalDateRange>>
+    ): Map<Long, List<Habit.CompletionRecord>>
 
     suspend fun insertCompletion(
         habitId: Long,