--- a/core/data/src/main/java/com/rendox/routinetracker/core/data/completion_history/CompletionHistoryRepositoryImpl.kt
+++ b/core/data/src/main/java/com/rendox/routinetracker/core/data/completion_history/CompletionHistoryRepositoryImpl.kt
@@ -9,12 +9,12 @@ class CompletionHistoryRepositoryImpl(
 ) : CompletionHistoryRepository {
 
     override suspend fun getRecordsInPeriod(
-        habitId: Long,
+        habit: Habit,
         minDate: LocalDate?,
         maxDate: LocalDate?,
     ): List<Habit.CompletionRecord> {
         return localDataSource.getRecordsInPeriod(
-            habitId, minDate, maxDate
+            habit, minDate, maxDate
         )
     }