--- a/core/data/src/main/java/com/rendox/routinetracker/core/data/habit/HabitRepository.kt
+++ b/core/data/src/main/java/com/rendox/routinetracker/core/data/habit/HabitRepository.kt
@@ -5,6 +5,7 @@ import com.rendox.routinetracker.core.model.Habit
 interface HabitRepository {
 
     suspend fun insertHabit(habit: Habit)
+    suspend fun insertHabits(habits: List<Habit>)
     suspend fun getHabitById(id: Long): Habit
     suspend fun getAllHabits(): List<Habit>
     suspend fun deleteHabit(id: Long)