--- a/core/data/src/main/java/com/rendox/routinetracker/core/data/vacation/VacationRepository.kt
+++ b/core/data/src/main/java/com/rendox/routinetracker/core/data/vacation/VacationRepository.kt
@@ -22,7 +22,7 @@ interface VacationRepository {
         maxDate: LocalDate? = null,
     ): List<Vacation>
 
-    suspend fun getAllVacations(): List<Pair<Long, Vacation>>
+    suspend fun getAllVacations(): Map<Long, List<Vacation>>
     suspend fun getLastVacation(habitId: Long): Vacation?
     suspend fun insertVacation(habitId: Long, vacation: Vacation)
     suspend fun deleteVacationById(id: Long)