--- a/core/data/src/main/java/com/rendox/routinetracker/core/data/vacation/VacationRepository.kt
+++ b/core/data/src/main/java/com/rendox/routinetracker/core/data/vacation/VacationRepository.kt
@@ -5,17 +5,6 @@ import kotlinx.datetime.LocalDate
 
 interface VacationRepository {
 
-    /** Returns a [Vacation] that includes [date]. */
-    suspend fun getVacationByDate(
-        habitId: Long, date: LocalDate
-    ): Vacation?
-
-    /** Returns a [Vacation] that was ended before the [currentDate]. */
-    suspend fun getPreviousVacation(
-        habitId: Long,
-        currentDate: LocalDate,
-    ): Vacation?
-
     suspend fun getVacationsInPeriod(
         habitId: Long,
         minDate: LocalDate? = null,
@@ -23,7 +12,6 @@ interface VacationRepository {
     ): List<Vacation>
 
     suspend fun getAllVacations(): Map<Long, List<Vacation>>
-    suspend fun getLastVacation(habitId: Long): Vacation?
     suspend fun insertVacation(habitId: Long, vacation: Vacation)
     suspend fun deleteVacationById(id: Long)
 }