--- a/core/data/src/main/java/com/rendox/routinetracker/core/data/vacation/VacationRepository.kt
+++ b/core/data/src/main/java/com/rendox/routinetracker/core/data/vacation/VacationRepository.kt
@@ -1,5 +1,6 @@
 package com.rendox.routinetracker.core.data.vacation
 
+import com.rendox.routinetracker.core.logic.time.LocalDateRange
 import com.rendox.routinetracker.core.model.Vacation
 import kotlinx.datetime.LocalDate
 
@@ -7,11 +8,15 @@ interface VacationRepository {
 
     suspend fun getVacationsInPeriod(
         habitId: Long,
-        minDate: LocalDate? = null,
-        maxDate: LocalDate? = null,
+        minDate: LocalDate,
+        maxDate: LocalDate,
     ): List<Vacation>
 
-    suspend fun getAllVacations(): Map<Long, List<Vacation>>
+    suspend fun getMultiHabitVacations(
+        habitsToPeriods: List<Pair<List<Long>, LocalDateRange>>,
+    ): Map<Long, List<Vacation>>
+
     suspend fun insertVacation(habitId: Long, vacation: Vacation)
+    suspend fun insertVacations(habitIdsToVacations: Map<Long, List<Vacation>>)
     suspend fun deleteVacationById(id: Long)
 }