--- a/core/data/src/main/java/com/rendox/routinetracker/core/data/vacation/VacationRepositoryImpl.kt
+++ b/core/data/src/main/java/com/rendox/routinetracker/core/data/vacation/VacationRepositoryImpl.kt
@@ -1,6 +1,7 @@
 package com.rendox.routinetracker.core.data.vacation
 
 import com.rendox.routinetracker.core.database.vacation.VacationLocalDataSource
+import com.rendox.routinetracker.core.logic.time.LocalDateRange
 import com.rendox.routinetracker.core.model.Vacation
 import kotlinx.datetime.LocalDate
 
@@ -10,18 +11,24 @@ class VacationRepositoryImpl(
 
     override suspend fun getVacationsInPeriod(
         habitId: Long,
-        minDate: LocalDate?,
-        maxDate: LocalDate?,
+        minDate: LocalDate,
+        maxDate: LocalDate,
     ): List<Vacation> {
         return localDataSource.getVacationsInPeriod(habitId, minDate, maxDate)
     }
 
-    override suspend fun getAllVacations(): Map<Long, List<Vacation>> {
-        return localDataSource.getAllVacations()
+    override suspend fun getMultiHabitVacations(
+        habitsToPeriods: List<Pair<List<Long>, LocalDateRange>>
+    ): Map<Long, List<Vacation>> {
+        return localDataSource.getMultiHabitVacations(habitsToPeriods)
     }
 
     override suspend fun insertVacation(habitId: Long, vacation: Vacation) {
         return localDataSource.insertVacation(habitId, vacation)
+    }
+
+    override suspend fun insertVacations(habitIdsToVacations: Map<Long, List<Vacation>>) {
+        return localDataSource.insertVacations(habitIdsToVacations)
     }
 
     override suspend fun deleteVacationById(id: Long) {