--- a/core/data/src/release/java/com/rendox/routinetracker/core/data/di/DatabasePrepopulatorModule.kt
+++ b/core/data/src/release/java/com/rendox/routinetracker/core/data/di/DatabasePrepopulatorModule.kt
@@ -1,0 +1,11 @@
+package com.rendox.routinetracker.core.data.di
+
+import com.rendox.routinetracker.core.data.DatabasePrepopulator
+import com.rendox.routinetracker.core.data.DatabasePrepopulatorNoOp
+import org.koin.dsl.module
+
+val databasePrepopulatorModule = module {
+    single<DatabasePrepopulator> {
+        DatabasePrepopulatorNoOp()
+    }
+}