--- a/core/database/src/main/java/com/rendox/routinetracker/core/database/completion_history/CompletionHistoryLocalDataSource.kt
+++ b/core/database/src/main/java/com/rendox/routinetracker/core/database/completion_history/CompletionHistoryLocalDataSource.kt
@@ -36,6 +36,10 @@ interface CompletionHistoryLocalDataSource {
         completionRecord: Habit.CompletionRecord,
     )
 
+    suspend fun insertCompletions(
+        completions: Map<Long, List<Habit.CompletionRecord>>
+    )
+
     suspend fun deleteCompletionByDate(
         habitId: Long,
         date: LocalDate,