--- a/core/database/src/main/java/com/rendox/routinetracker/core/database/completion_history/CompletionHistoryLocalDataSource.kt
+++ b/core/database/src/main/java/com/rendox/routinetracker/core/database/completion_history/CompletionHistoryLocalDataSource.kt
@@ -29,7 +29,7 @@ interface CompletionHistoryLocalDataSource {
         maxDate: LocalDate?,
     ): List<Habit.CompletionRecord>
 
-    suspend fun getAllRecords(): List<Pair<Long, Habit.CompletionRecord>>
+    suspend fun getAllRecords(): Map<Long, List<Habit.CompletionRecord>>
 
     suspend fun insertCompletion(
         habitId: Long,