Restructured the pipeline of agenda habits loading resulting in ~50% faster retrieval.

The performance improvement comes from a significant reduction in IO operations. Previously, habits were loaded one by one, **each** taking three requests to the database. Now, database requests return collections so it only takes three requests **in total** to load the agenda. These three include retrieving a list of habits, a multimap of completion records, and a multimap of vacation records.

This PR also includes optimizations for a large data set. Agenda loading stays relatively fast even when there are dozens of habits, completions, and vacations in the database. This is achieved by limiting number of the retrieved entries. 

- We limit the number of habits to those that are started and aren't yet finished at the moment.
- We limit the number of completions and vacations to only those that are within the current period.

Since retrieving histories for every habit individually is no longer possible, we now group habits by their schedule, determine the common period for the whole group, and retrieve their histories in one query. This is performant because habits with the same schedule have a very close period.

https://github.com/DanielRendox/RoutineTracker/assets/107246216/ee187167-f691-49a0-822e-ac1a3491dec8


|**Measurements**| **Before** | **After** | **Improvement** |
| --------------------| ----------- | ------- | ---------------------- |
| 50 habits | 1067 ms | 214 ms | 80% |
| 15 habits | 463 ms | 217 ms | 53% |

- hundreds of completion entries in db in each example
- measurements made in release mode on a real device 

Before the introduced optimizations the operation of retrieving agenda had linear complexity — the more habits were stored, the longer it took to compute them. With the improvements made, it got close to the logarithmic complexity.