@CodiumAI-Agent 2024-06-29
**Changelog updates:** 🔄

## 2024-06-29

### Changed
- Optimized loading habits on the Agenda Screen, resulting in ~40% faster retrieval.
- Reduced the number of database requests from three per habit to three in total.
- Limited the number of retrieved entries for large data sets to maintain performance.
- Grouped habits by schedule for efficient history retrieval.


>to commit the new content to the CHANGELOG.md file, please type:
>'/update_changelog --pr_update_changelog.push_changelog_changes=true'